# openapi.yml
components:
  schemas:
...
    Current:
      allOf:
      - $ref: '#/components/schemas/TimeRef'
      - properties:
          song:
            $ref: '#/components/schemas/IdOrRep_Song'
        required:
        - song
        type: object
...
    IdOrRep_QueueConfig:
      description: Either an id or a representation of a resource
      oneOf:
      - format: uuid
        type: string
      - properties:
          autolike:
            type: boolean
          banned_role:
            type: string
          created:
            format: date-time
            type: string
          creator_role:
            type: string
          default_role:
            type: string
          id:
            format: uuid
            type: string
          roles:
            additionalProperties:
              $ref: '#/components/schemas/QueueUserRole'
            propertyNames:
              type: string
            type: object
          updated:
            format: date-time
            type: string
        required:
        - id
        - creator_role
        - default_role
        - banned_role
        - roles
        - autolike
        - created
        - updated
        type: object
    IdOrRep_Song:
      description: Either an id or a representation of a resource
      oneOf:
      - format: uuid
        type: string
      - properties:
          added_by:
            description: User that first added the song
            format: uuid
            type:
            - string
            - 'null'
          created:
            description: When the song was added
            format: date-time
            type: string
          details:
            description: 'Additional data from the song source


              These contains stuff like thumbnails,

              the artist, an url or any source specific

              data. They are provided on-demand

              as they require querying the source service'
            type:
            - object
            - 'null'
          duration:
            description: Duration of the song
            type: string
          id:
            description: Unique id of the song
            format: uuid
            type: string
          source:
            description: Source that provided this song
            type: string
          title:
            description: Title of the song
            type: string
        required:
        - id
        - title
        - duration
        - created
        - source
        type: object
...
    Queue:
      properties:
        code:
          type: string
        config:
          $ref: '#/components/schemas/IdOrRep_QueueConfig'
        created:
          format: date-time
          type: string
        current:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Current'
        id:
          format: uuid
          type: string
        player_state_id:
          format: uuid
          type: string
        queue:
          additionalProperties:
            $ref: '#/components/schemas/QueuedSong'
          propertyNames:
            format: uuid
            type: string
          type: object
        updated:
          format: date-time
          type: string
        user:
          $ref: '#/components/schemas/QueueUserDto'
      required:
      - id
      - code
      - player_state_id
      - config
      - user
      - queue
      - created
      - updated
      type: object
...
    QueueUserAction:
      description: An action a user makes on a queue
      enum:
      - GET_QUEUE
      - DELETE_QUEUE
      - CONFIGURE_QUEUE
      - REMOVE_SONG
      - BAN_SONG
      - UNBAN_SONG
      - ENQUEUE_SONG
      - PLAY_SONG
      - PAUSE_SONG
      - NEXT_SONG
      - AUTO_NEXT_SONG
      - LIKE_SONG
      - BAN_USER
      - UNBAN_USER
      - REMOVE_USER
      type: string
    QueueUserDto:
      properties:
        auto_like:
          type: boolean
        likes:
          format: int32
          minimum: 0
          type: integer
        role:
          type: string
      required:
      - auto_like
      - likes
      - role
      type: object
    QueueUserRole:
      properties:
        can_grant:
          items:
            type: string
          type: array
          uniqueItems: true
        can_revoke:
          items:
            type: string
          type: array
          uniqueItems: true
        id:
          format: uuid
          type: string
        max_likes:
          format: int32
          minimum: 0
          type: integer
        permissions:
          items:
            $ref: '#/components/schemas/QueueUserAction'
          type: array
          uniqueItems: true
      required:
      - id
      - max_likes
      - permissions
      - can_grant
      - can_revoke
      type: object
...
    QueuedSong:
      properties:
        likes:
          description: Number of likes this song has
          format: int32
          minimum: 0
          type: integer
        queued_at:
          description: When the song was queued
          format: date-time
          type: string
        song:
          $ref: '#/components/schemas/IdOrRep_Song'
          description: Song that was queued
        user_likes:
          description: Number of likes this song has by the current user
          format: int32
          minimum: 0
          type: integer
      required:
      - song
      - queued_at
      - likes
      - user_likes
      type: object
...
    TimeRef:
      oneOf:
      - properties:
          position:
            type: string
        required:
        - position
        type: object
      - properties:
          starts_at:
            format: date-time
            type: string
        required:
        - starts_at
        type: object
...
info:
  contact:
    email: zannabianca199712@gmail.com
    name: zannabianca1997
  description: '> A communist music queue


    `apelle` is a service for handling a shared music queue. Users can insert songs

    in the queues, and upvote them to push them upward. `apelle` will track the

    position of each song in the queue, and the position of the currently playing

    song.


    This is the documentation for the external API.

    '
  license:
    identifier: MIT
    name: MIT
  summary: Apelle public API
  title: Apelle API
  version: 2.0.0-alpha
openapi: 3.1.0
paths:
...
  /api/queues/{queue_id}:
    delete:
      operationId: queues_delete
      parameters:
      - in: path
        name: queue_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '204':
          description: Queue deleted
        '403':
          description: User is not allowed to delete the queue
        '500':
          description: Internal server error
      summary: Read the queue data
      tags:
      - service-queues
    get:
      operationId: queues_get
      parameters:
      - description: Return the full queue config instead of just the UUID
        in: query
        name: config
        required: false
        schema:
          type: boolean
      - description: Return the full song data instead of just the UUID
        in: query
        name: songs
        required: false
        schema:
          type: boolean
      - in: query
        name: details
        required: false
        schema:
          type: boolean
      - in: path
        name: queue_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
          description: Queue created
        '403':
          description: User is not allowed to read the queue data
        '500':
          description: Internal server error
        '502':
          description: Error returned from songs service
      summary: Read the queue data
      tags:
      - service-queues
...